разделяет код на модули и добавляет сборщик
