+++ .gitignore
@@ -12,3 +12,4 @@ Thumbs.db
 node_modules/
 
 # Файлы сборки (допишите самостоятельно)
+public/bundle.js
