+++ src/components/loadMoreButton.js
@@ -1,0 +1,7 @@
+const getLoadMoreButtonMarkup = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {getLoadMoreButtonMarkup};
