+++ webpack.config.js
@@ -1,0 +1,15 @@
+const path = require(`path`);
+
+module.exports = {
+  mode: "development",
+  entry: "./src/main.js",
+  output: {
+    filename: "bundle.js",
+    path: path.resolve(__dirname, "public"),
+  },
+  devtool: "source-map",
+  devServer: {
+    contentBase: path.join(__dirname, 'public'),
+    watchContentBase: true
+  },
+};
